Removed Support Related Controllers as they are not used anymore (Drift replaced support system) Converted most code to be ES6 (and have Babel transpile it to basic JS) Caveat: Controller loading can't be ES6 since it causes an issue with how angularjs, so that will remain to use function instead of directly using the variable of the controller. Reformatted most code so that its cleaner and removes excess whitespace
Note: Not sure if there is any caveats to converting the code to ES6 using a converter, but I guess we will learn about them soon haha

For some reason, Gulp/Webpack now compress the code to less than 1.3mb js file, so thats great. Possibly even less with gzip. Which is even better. May just drop caching for a while since this should be acceptable.